backend/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.2.0
Django==5.2.6
djangorestframework==3.16.1
drf-yasg==1.21.11
h11==0.16.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
simplejson==3.20.2
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
